+++ sdk/io/asn.1/basic/visitor/asn1_visitor.hpp
@@ -19,15 +19,22 @@
 namespace hotplace {
 namespace io {
 
+/**
+ * @brief   visitor
+ * @example
+ *          auto lambda = [this](asn1_object* item) -> void { dosomething(); };
+ *          asn1_visitor visitor(&runtime, lambda);
+ *          visitor.visit(schema);
+ */
 class asn1_visitor {
    public:
-    asn1_visitor(asn1_runtime* runtime, std::function<void(asn1_object*)> func);
+    asn1_visitor(const asn1_runtime* runtime, std::function<void(asn1_object*)> func);
     virtual ~asn1_visitor() = default;
 
     virtual void visit(asn1_object* object);
 
    private:
-    asn1_runtime* _runtime;
+    const asn1_runtime* _runtime;
     std::function<void(asn1_object*)> _func;
 };