+++ CHANGELOG.md
@@ -35,6 +35,10 @@
     * CBOR
 
 * details
+  * Revision 1083
+    * string_set - erase_range, intersect
+    * graph - topological sort
+    * asn1_runtime - resolve, is_resolved
   * Revision 1082
     * asn1_publisher - Constraints
       * exclusive boundary (1..<256, 1<..256, 1<..<256)